% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations to carry out a binomial experiment of a given number of trials and success probability}

\item{n}{The number of trials of a given experiment}

\item{p}{The probability of success on a given trial}
}
\value{
A table of relative frequencies of successes and a barplot of relative frequencies of successes across all iterations
}
\description{
Binomial Simulation
}
\examples{
mybin(iter=100,n=10, p=0.5)
}
