% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x)
}
\arguments{
\item{x}{Vector containing sample data}
}
\value{
95% confidence interval of the population mean of x
}
\description{
myci
}
\examples{
myci(c(1,2,3))
}
