% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Log likelihood function to be applied to pairs of data values and parameter values}

\item{x}{Vector containing data}

\item{param}{Vector containing parameter values}

\item{...}{Additional parameters uesd in the development of plots}
}
\value{
List containing the index i where the max parameter value occurs, the maximum likelihood estimate parami, the max log likelihood value yi, and the slopes of the log likelihood around the maximum
}
\description{
mymaxlik
}
\examples{
\dontrun{mymaxlik(lfun=logbin,x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000))}
}
