% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the distribution}

\item{sigma}{Standard deviation of the distribution}

\item{a}{Upper bound of lower tail probability}
}
\value{
A list of the parameters
}
\description{
Normal Curve
}
\examples{
myncurve(5,3,2)
}
