% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{T-calc value obtained by assuming the null hypothesis is true and plugging in the null value into the T-statistic equation}

\item{xmax}{X-value used to create the bounds for the curve plotted for the T-statistic distribution}

\item{n}{Sample size}

\item{alpha}{Value that determines the confidence level of the confidence interval and used to compare with the p-value to determine whether to reject the null hypothesis or not}
}
\value{
A list containing the quantile t_(alpha/2) as q and the area corresponding to the p-value as pvalue
}
\description{
mypvalue
}
\examples{
mypvalue(t0=1.326252)
}
