% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats on the flight}

\item{gamma}{Probability of overbooking}

\item{p}{Probability of success (a person showing up)}
}
\value{
A list of the optimal number of tickets to sell for the discrete distribution and continuous approximation, the number of seats on the flight, the probability of overbooking, and the probability of a person showing up
}
\description{
ntickets
}
\examples{
ntickets(400,0.02,0.95)
}
